# api/src/swagger.json
{
    "openapi": "3.0.0",
    "servers": [
      {
        "url": "http://example.com",
        "description": "",
        "variables": {}
      }
    ],
    "info": {
      "version": "aa6a9ff1-052f-4deb-b68c-4b06dabe0335",
      "title": "challenger-for-software-engeneer-position",
      "description": "",
      "termsOfService": "",
      "contact": {},
      "license": {
        "name": ""
      }
    },
    "paths": {
...
      "/user/{id}": {
        "put": {
          "summary": "Update user",
          "description": "Update user",
          "operationId": "Updateuser",
          "parameters": [
            {
              "$ref": "#/components/parameters/id"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "tags": [
            "user"
          ]
        },
        "get": {
          "summary": "Get user by id",
          "description": "Get user by id",
          "operationId": "Getuserbyid",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "tags": [
            "user"
          ]
        },
        "delete": {
          "summary": "Delete user",
          "description": "Delete user",
          "operationId": "Deleteuser",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "tags": [
            "user"
          ]
        }
      },
...
    },
    "components": {
      "parameters": {
        "id": {
          "name": "id",
          "in": "path",
          "required": true,
          "style": "simple",
          "schema": {
            "type": "number"
          }
        }
      }
    },
    "security": [],
...
  }